// geoclient/lib/api/generated/openapi.d.ts
export interface components {
    schemas: {
// ...
        PasswordChangeIn: {
            /**
             * Old Password
             * @description 原密码
             */
            old_password: string;
            /**
             * New Password
             * @description 新密码，至少6位
             */
            new_password: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}