// geoclient/lib/api/generated/user-openapi.d.ts
export interface components {
    schemas: {
// ...
        "admin.v1.AdminLoginReply": {
            /** @description 访问令牌。 */
            accessToken?: string;
            /** @description 刷新令牌。 */
            refreshToken?: string;
            /**
             * Format: date-time
             * @description 访问令牌过期时间。
             */
            accessExpiresAt?: string;
            /** @description 平台管理员。 */
            admin?: components["schemas"]["admin.v1.AdminProfile"];
        };
        /** @description 平台管理员登录请求参数。 */
        "admin.v1.AdminLoginRequest": {
            /** @description 登录用户名。 */
            username?: string;
            /** @description 登录密码。 */
            password?: string;
            /** @description 客户端设备编号。 */
            deviceId?: string;
        };
// ...
        "user.v1.LoginReply": {
            /** @description 访问令牌。 */
            accessToken?: string;
            /** @description 刷新令牌。 */
            refreshToken?: string;
            /**
             * Format: date-time
             * @description 访问令牌过期时间。
             */
            accessExpiresAt?: string;
            /** @description 企业。 */
            enterprise?: components["schemas"]["user.v1.EnterpriseProfile"];
        };
        /** @description 登录请求参数。 */
        "user.v1.LoginRequest": {
            /** @description 登录用户名。 */
            username?: string;
            /** @description 登录密码。 */
            password?: string;
            /** @description 客户端设备编号。 */
            deviceId?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}